Fix LMDB map size multiplication width
The database open path multiplied the unsigned int db_max_size value before promotion to size_t. Manual limits of 4096 MiB or more could therefore wrap before mdb_env_set_mapsize received the result.

Promote db_max_size to size_t before multiplying by the MiB constant, matching the other LMDB resize paths and preserving large configured map sizes.

+++ src/library/database.c
@@ -2551,7 +2551,8 @@ static int init_db_with_generations(const conf_t *config,
 		return 2;
 	}
 
-	if (mdb_env_set_mapsize(env, config->db_max_size*MEGABYTE)) {
+	if (mdb_env_set_mapsize(env,
+			(size_t)config->db_max_size * MEGABYTE)) {
 		/* Clean up environment on failure */
 		mdb_env_close(env);
 		env = NULL;
